V2.1: Add Key class
Add a class that represents a key (See documentation of Key class for
more details). The Key class will, of course, be used in the
implementation of the addKey method. As stated in the previous commit,
this is done to make the application more scalable and easier to
change/customize by making it more dynamic.